+++ src/main/java/com/policy/bazaar/globalresponse/Response.java
@@ -3,14 +3,14 @@ package com.policy.bazaar.globalresponse;
 public class Response {
 
 	
-	  private Integer userType;
+	  private Byte userType;
 	  private Boolean Auth = true;
 	
 	  
-	  public Integer getUserType() {
+	  public Byte getUserType() {
 		return userType;
 	}
-	public void setUserType(Integer userType) {
+	public void setUserType(Byte userType) {
 		this.userType = userType;
 	}
 	public Boolean getAuth() {
